fix: resolve all remaining QA audit items
Previously skipped items, now fixed:

M5:  Frontend middleware for dashboard route protection. Login
     sets a kolamba_auth cookie; middleware redirects to /login
     if cookie is absent. Added setAuthTokens/clearAuthTokens
     helpers in api.ts for consistent token management.

M8:  Fixed Decimal/float type mismatch — Community and
     ArtistTourDate lat/lng annotations now use Mapped[Decimal]
     to match the Numeric column type.

M10: Geocoding rate limiting — async lock enforces max 1 req/sec
     to Nominatim API per their usage policy.

M11: Kept as-is — dashboard pages need user data (community_id,
     artist_id) for queries, not just auth. Separate /auth/me
     call is necessary.

L6:  Favorites now fetched in parallel with Promise.all instead
     of sequential N+1 loop.

L7:  Swagger docs, ReDoc, and OpenAPI schema disabled in
     production (only available when env=development).

L8:  Search endpoint falls back to LIKE-only queries when running
     on SQLite (for tests), using PostgreSQL full-text search only
     when on PostgreSQL.

L9:  Addressed by M5 middleware — unauthenticated users are
     redirected before dashboard pages load.

+++ backend/app/models/artist_tour_date.py
@@ -1,6 +1,7 @@
 """ArtistTourDate model - tour dates announced by artists."""
 
 from datetime import datetime, date, timezone
+from decimal import Decimal
 from typing import TYPE_CHECKING, Optional
 from sqlalchemy import String, Integer, Boolean, DateTime, Date, Text, ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -25,8 +26,8 @@ class ArtistTourDate(Base):
 
     # Location info
     location: Mapped[str] = mapped_column(String(255), nullable=False)
-    latitude: Mapped[Optional[float]] = mapped_column(Numeric(10, 8), nullable=True)
-    longitude: Mapped[Optional[float]] = mapped_column(Numeric(11, 8), nullable=True)
+    latitude: Mapped[Optional[Decimal]] = mapped_column(Numeric(10, 8), nullable=True)
+    longitude: Mapped[Optional[Decimal]] = mapped_column(Numeric(11, 8), nullable=True)
 
     # Dates
     start_date: Mapped[date] = mapped_column(Date, nullable=False)
